Update broken_travel.py
There was different quotations for the input statement that needed to be corrected. After the int, they kept a dot instead of a bracket. In the if statement there was no semicolon. In the elif statement you can use the built in function and to join the two possibilities. At the end they wrote an elif statement instead of an else statement.